+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@libsql/client": "^0.6.0",
     "express": "4.19.2",
+    "js-cookie": "^3.0.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "6.23.0",
