+++ src/components/header/SessionButton.jsx
@@ -1,24 +1,18 @@
 import {UserIcon} from "../../assets/icons/UserIcon"
+import { useUserStore } from "../../store/user"
+import { LoginOption } from "./sessionOptions/LoginOption"
+import { LogoutOption } from "./sessionOptions/LogoutOption"
+
 
 export const SessionButton = ()=>{
 
-    const handleClick = ()=>{
-        window.location.href='/session'
-        
-    }
+    const {isLogged} = useUserStore()
 
     return(
-        <li class="flex-grow flex justify-center items-end">  
-            <a class="flex items-end" href="javascript:void(0)" onClick={handleClick}>
-            
-                <UserIcon/>
-            
-                <div>
-                    <p>Bienvenido!</p>
-                    <b>Inicia sesión / Regístrate</b>    
-                </div>
-                
-            </a>
-        </li>
+        isLogged?(
+            <LogoutOption/>
+        ):(
+            <LoginOption/>
+        )
     )
 }
